Alerts: refactor id types

+++ mtTorrent/Core/Api/CoreImpl.cpp
@@ -67,7 +67,7 @@ mtt::Status mttApi::Core::removeTorrent(TorrentPtr torrent, bool deleteFiles)
 	return static_cast<mtt::Core*>(this)->removeTorrent(torrent->getFileInfo().info.hash, deleteFiles);
 }
 
-void mttApi::Core::registerAlerts(uint32_t alertMask)
+void mttApi::Core::registerAlerts(uint64_t alertMask)
 {
 	mtt::AlertsManager::Get().registerAlerts(alertMask);
 }
